Screen/Linux: Refine Prime Optimus support for NVidia blob.
Use presence of a special XAtom set by our customized
modesetting-ddx to signal the custom ddx is active.

If so, enable our custom timestamping and scheduling and
signal it to the user. If not, warn user about expected
bad timing and advise the user to install the custom
modesetting ddx.

Additionally on hybrid graphics setups with Intel iGPU,
tell user to upgrade to at least Linux 4.5 to get proper
prime sync support if an older kernel is active.