Screen/Linux: Make use of experimental low-lag optimizations for Prime.
If the X-Atom "PrimeTimingHack2" gets detected then PTB is running on
a XOrg 1.19 modesetting-ddx with additional low-lag optimizations.

In that case the 1-frame extra-lag of outputSource->outputSlave
Prime mode should usually not exist anymore, so we don't need
to compensate for 1 frame lag in swap scheduling and can always
wait for true flip completion packets instead of waiting for
"flip scheduled" packets. This should bring performance mostly
in line with a non-Optimus setup - at least in a perfect world...

In reality this may need more tweaking, and some tests indicate
that the low-lag optimization in the modesetting-ddx may not
work without bad side effects, like massive stalls or long hangs
of the GUI on multi-display setups. This may make these optimizations
unworkable. But add the code to PTB anyway, just in case i manage
to get the ddx code working properly.