Screen/Linux: Handle Intel iGPU + AMD dGPU under amdgpu-kms.
If DRI3 Prime renderoffload is used on a Intel + AMD hybrid
graphics laptop with a AMD gpu that is driven by amdgpu-kms
then check if the Linux kernel is at least Linux 4.9.

Earlier kernels do not provide proper dmabuf-fence sync between
intel-kms and amdgpu-kms, so visual stimuli can get visually
corrupted under high graphics loads.

Fixed in Linux 4.9-rc6, ergo in Linux 4.9 (ETA early December 2016).
A backport to Linux 4.8 probably in the December/January timeframe
is somewhat likely, so users of Ubuntu 16.10 / Ubuntu 16.04.2 LTS
should be good in the foreseeable future. Until this has happened,
we output a warning and recommendation to upgrade to 4.9+ on older
kernels.

-> Screen.mex for 64-Bit Octave on Linux rebuilt.