Screen/Linux: Update Linux kernel requirements for PRIME again.
On Intel iGPU require Linux 4.6+ instead of 4.5+, as NVidia
found some performance issues with concurrent flips and hw
cursor updates on their proprietary driver under Optimus.
Issues are present in 4.5, but apparently fixed in 4.6+.

On Intel iGPU + AMD dgpu with amdgpu-kms, the dmabuf-fence
fix has been backported for upcoming Linux 4.8.11, so require
that as minimum instead of Linux 4.9. Unfortunately i don't
know yet the timeline for a backport to Ubuntu's 4.8-based
kernel.

-> Screen.mex for 64-Bit Octave on Linux rebuilt.