MOGL: Fix glClearNamedFramebufferfi()
Implement it (manually now in gl_manual.c) to conform to the
bug-fixed OpenGL 4.5 spec, so it works with current GLEW-2.0
and should work with bug-fixed OpenGL drivers, e.g., Mesa 12+.

Downside is it would malfunction or even crash now on drivers
which still implement the broken 4.5 spec, e.g., Mesa < 12 on
*buntu 16.04.1/2 - LTS, and on various not-up-to-date Windows
OpenGL drivers. As this dilemma is not resolvable we add a
one-time warning to glClearNamedFramebufferfi.m that warns user
about this peculiarity/hazard and recommends a driver upgrade
to avoid mayhem.

Fixes #372