MOGL: Fix glA* to glO* for Matlab R2015b+
This probably fixes letters A to O of the alphabet of glXXX functions.

Letters P-Z obviously tbd.

So it turns out the failure of many MOGL functions which return data
into preallocated buffer is not due to Mathworks breaking mex api
backwards compatibility. Nope, no such thing. Instead it is simply a
bug or quirk introduced in R2015b and later: Whenever the interpreter
has to pass only a scalar value ie. 1 element, it does so by passing
essentially a private copy, which can be read by moglcore, but not
refilled with new data. This in accordance to our old assumption.
However, as soon as the passed in data is a true vector or matrix with
more than 1 element, R2015b’s behaviour reverts back to the way things
used to be. We can pass data back to Matlab M-Files in prhs
right-hand-side arguments!

To take advantage of this, we just have to make sure to always pass in
at least a 2 element vector. Whenever that can’t be guaranteed, we
modify the M-File wrapper to artificially allocate a vector with 1
extra element (ergo at least 1 true + 1 extra = 2 elements), so the mex
api lets us pass data back. Then we truncate the mex-returned vector by
one element before passing it back to our calling function.

This trick worked with glGenTextures and glGenBuffers, as tested on
R2015b on both Linux and OSX, so should work with other functions as
well.

-> Still to do: Convert glGetPxxxx() - glGetZxxxx() functions.
-> See issue #308