PsychHID/OSX: Handle MacBook"Pro" 2016 brain-damage.
Find default keyboard device on OSX by starting with the fail-safe
assumption that the first detected keyboard is really a keyboard, then
iterate over all candidates, first keyboards, then keypads, mice,
pointing devices, joysticks, gamepads. If none is valid, we will fall
back to device 0 as best "worst choice”, the same choice we used
successfully for the last 12+ years.

In the good old past we simply used device 0, the first keyboard, but
the iToys company decided to make life difficult again when introducing
the MacBook"Pro" 2016 with a new iToy gadget, the oh so magic Touchbar.
This Touchbar enumerates as the first "keyboard" in the system, with no
transport, version, manufacturer name, product name or serial number,
but instead with 1046 keys (yay!). Luckily this thing also has an
invalid locationID of zero, so we can filter it out with this code by
only accepting keyboards with a non-zero locationID. The 2nd keyboard
is then the actual keyboard of the MacBook"Pro" 2016.