datapixxmakemex: Fix build for 64-Bit Matlab on OSX.
In the line with the linker parameters, replace “” double-quotes by 2
successive single-quotes ‘’ and remove all previously existing
successive single-quotes ’’ in that line without replacement. This way
the whole linker options string is enclosed in a pair of 2 successive
single quotes: ’’

This fixes the build of Datapixx on R2012a on OSX 10.12.