OSX/Octave: Upgrade build scripts to Octave-4.2 native build.
Now we use Octave-4.2, built from source via HomeBrew, to build mex
files for Octave-4.2 on OSX, using the 10.12 Sierra SDK, for OSX 10.11
and 10.12.
Because of new OSX 10.12 linker brain-damage, we can’t use
octave_config_info.octlibdir anymore for the base path to strip from
mex files for our rpath hack. Although octave_config_info returns a lib
path to /usr/local/Cellar/octave/… and the compiler gets exactly that
library path passed as -L path, apparently the brain-damaged OS X inker
decides to instead encode a /usr/local/opt/octave/… path into the mex
files. That path is an alias - a symlink - to the real absolute
/usr/local/Cellar path, but for god knows what reason, this is what the
linker does. So know we have to hard-code the path for a HomeBrew
install of Octave for this to work. Thanks Apple for making life
miserable at the end of the year as well.