moglmorpher: Fix rendering of single mesh with vertex colors.
If a vertex color vector was used instead of, or additionally
to, textures, and only 1 single mesh was added with 'addMesh'
then the VBO with the vertex colors did not get initialized
with the passed in vertex colors. A mistake in resource allocation
only caused allocation of a VBO, but not actual init with data, so
the data inside the VBO was undefined. Fix this special case by
passing the moglsingle(ctx.vertexcolors) to the glBufferData()
one time call, instead of a NULL pointer.

Also white-space/indentation/style cleanup across the file.