Mainly adding LsfToPsf and PsfToLsf, but also cleaning up comments on some other routines.
I put the test routine PsychOpticsTest in this directory rather than in PsychTests.  At various times we have wondered whether tests should go with the routines they test rather than in a single directory.  There are advantages to each scheme.  In this case, I sometimes share this whole folder with people who don’t otherwise need the PTB and having the test here is good for that.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>