Screen/Linux: Fix GetMouse() for multi-x-screen window local coordinates.
We didn't pass in the actual onscreen window x-Window handle and
did not return window local coordinates, but always X-Screen
RootWindow relative coordinates.

This worked fine on single-x-screen setups, but failed on
multi-x-screen setups if a windowHandle was given to retrieve
mouse pointer position relative to the onscreen window origin.

Fix it by modifying Screen('GetMousehelper') to return window relative
coordinates, for the given onscreen window if a onscreen window handle
is given, or for the RootWindow of an X-Screen if a screenId is given,
or for the RootWindow of X-Screen 0 if nothing is provided.

Also add special case to GetMouse.m for Linux + onscreen window handle
to return mouse position without further remapping. This should also
be the right thing to do for Wayland.

-> Screen for 64-Bit Octave on Linux rebuilt for testing.