Screen/Linux: Fix GetMouse harder on multi-X-Screens.
The pointer query functions return (0,0) if the mouse pointer
is not on the X-Screen which we query, or where our given
onscreen window is located. That's unfortunate as (0,0)
does not provide meaningful info if GetMouse(screenNumber)
is called, and even less useful info if GetMouse(window)
is called, e.g., if one wants to find out if the mouse pointer
is inside the window, that would suggest the pointer is inside
the window (at top-left 0,0 window origin) if the pointer is
not even on the same screen as the window!

Our GetMouse API is not expressive enough to convey the
subtleties of a multi-X-Screen Linux setup properly to
scripts without breaking backwards compatibility, so we
need to fudge something hopefully good and backwards compatible
enough:

Iff the mouse pointer is not even on the X-Screen for which
we want to return relative mouse position, or for which we
want to return mouse position relative to a given onscreen
window, fake it as follows:

If an onscreen window handle was provided, simply return
(mx,my) = (-1,-1). No useful info about real pointer location,
wrt. the onscreen window, but at least a hit test if a mouse
pointer is inside the window would properly fail this way as
expected, ie. report in "not inside window" ie. not inside
the windows local rectangle, ie. IsInRect(mx,my, Screen('Rect', w))
for [mx,my] = GetMouse(w) would return false.

If a screen number was provided instead, just return pointer
coordinates relative to the X-Screen's root window, like we did
up to now, better than nothing. This is backwards compatible to
existing scripts, or to scripts which just call GetMouse() - a
call that translates to GetMouse(0) for X-Screen 0. Note that
the coordinates returned are not actually relative to the root
window of the queried screen, but simply to the root window on which
the mouse is actually positioned! This was always the case and is
kind of odd, but apparently the XQueryPointer and XIQueryPointer
simply don't know the overall desktop layout of a display connection,
so can not return proper root window relative coordinates if the
pointer is not actually inside the given root window. Anyway, it
was good enough in the past and during all my testing, so this has
to do.