VRNextPrep: PsychImaging - Different OpenWindow setup for new VR runtimes.
On current VR runtimes like Oculus 1.x, OpenVR etc., we don't do the
undistortion
rendering ourselves, but submit (ourselves and the frames) to a VR
compositor.

Therefore our current design does not use panel-fitting and similar
stuff to get
the to-be-submitted renderbuffers to the proper scale, but instead we
select
the redirected output mode and set the size of the parent onscreen
window to
the maximum recommended size of the per-eye output buffers.

Add corresponding setup code to PsychImaging OpenWindow for our first
test
candidate, the future PsychOculusVR1 driver for OculusVR 1.x runtimes.
Leave the
old code in place for the legacy PsychOculusVR driver for Oculus 0.5
runtimes.

-> This is still a very clumsy way of setting pipeline buffer
dimensions! May need to
refine this with some dedicated setup flag or such, to decouple the true
size of the
onscreen window from the size of the buffers for the VR compositor. The
onscreen
window is usually not useful for anything under such "Direct output
mode" runtimes,
but can still have value during development to show debug output.