VRNextPrep: PsychImaging - More setup code for the PsychOculusVR1 driver.
When an onscreen window is set up for use as parent window for a HMD
driven by the PsychOculusVR1 driver then use 'SkipSyncTests' preference
setting, as we are not interested in the display timing of the onscreen
window - it is either totally inactive, or only used to display
debug/diagnostic output during development/debugging of a VR script -
but in the display timing of the HMD. The HMD needs totally different
timestamping mechanisms than our standard ones, so sync tests and
calibrations don't meaningfully apply.

Additionally, select 'OpenWindow' specialFlags
kPsychGUIWindowWMPositioned, so the onscreen window gets treated like a
normal GUI window. This allows the WM and user to move it around, resize
it, and most importantly move it out of the way if it isn't needed to
display debug output.