VRHMDDemo: Adapt rendering from one-shot to rerender in a loop.
Our new zero-copy optimizations for use with current VR runtimes (at
least the Oculus 1.0 runtime) do not mingle well with a 'Flip' that does
not not clear the framebuffer. Therefore sacrifice that little "do
absolutely nothing" benchmark loop.

Also now render static image in a loop, as current VR compositors don't
like not getting regular new frames submitted. May need to deal with
this in some more clever way, but for now we are just nice to them.