Screen: Add optional 'fbOverrideRect' parameter to 'OpenWindow'
'fbOverrideRect' allows to override the size used for the
onscreen windows framebuffer, ie. it overrides windowRecord->rect.

Normally the framebuffer size is given by the 'rect' parameter
and corresponds to the true size of the onscreen window on the
display, and thereby its true system framebuffer size.

With this override one can decouple the true size of the window
on the windowing system display from the size with which all
internal rendering and image processing operates, iow. the
imaging pipeline can assume a different size of the raw framebuffer
from the real size.

This is mostly useful when the primary target for display is not
the actual onscreen window "inside" the operating systems windowing
system, but some 3rd party image consumer, e.g., some special output
device.

The first user of this facility are VR runtimes/compositors
which display the rendered content onto some VR-HMD which is not
part of the regular desktop GUI / windowing system, but instead
displays via some special "Direct output mode" facility.
Here the actual onscreen window only acts as host/parent to
provide an OpenGL rendering context for stimulus rendering and
post-processing, but final content is then sent somewhere else.
The onscreen window is either not used at all, and possibly hidden,
or only for some debug output.