Screen: Make sure expected onset time prediction never goes into past.
Iff an immediate flip ('flipWhen' == 0) is requested and
flip timestamping returns stimulus onset time (=end of vblank)
as vblank start, and a full 'return from flip' -> Render -> 'Flip'
cycle manages to execute within the vblank interval of last completed
flip, then it could have happened that ...

tshouldflip < windowRecord->time_at_last_vbl

... which would have let to a setting of expected tshouldflip
1 refresh in the past, and finally to a expected onset time
of most recent vblank. As that is impossible to achieve, it
led to a false positive for a missed flip deadline.

Fix this by making sure that tshouldflip is never in the
past.

-> Found during test with VBLSyncTest([],0); on AMD R9 380 under
amdgpu-pro.