psychtoolbox.rules: Add rule for the Photo Research PR-670 photometer et al.
If a product from Photo Research Inc. is detected, make sure that
the ModemManager daemon completely ignores it. Otherwise it would
try to access the device as a modem, which would fail, but would
confuse at least the PR-670 photometer and prevent it entering
remote control mode later when a PTB script calls PR670init.

-> Hurray to the genius that thought exposing a photometer as a
   standard modem was a bright idea. Especially if the photometer
   is not able to deal with then getting accessed as a modem.
   One doesn't need a photometer to see this idea wasn't the sharpest.

-> Btw. the same problem happens on OSX 10.11 El Capitan, but not
   on OS 10.12 Sierra. Maybe some similar root cause in Apples world?

-> PR-670 now verified to work on Linux.