Screen: Add new variables which can be passed back to M-Functions from within imaging pipeline.
Now the following new variabes names can be specified in a runtime
function (M-Function / expression) as part of a imaging pipeline
processin slot. They will get appropriate values assigned in the
callers workspace, iow. represent local variables:

IMAGINGPIPE_FLIPTWHEN = The 'when' target stimulus onset time passed into
Screen('Flip/AsyncFlipBegin', window, when); or zero if none was specified,
ie. immediate flip requested, or -DBL_MAX if unavailable.

IMAGINGPIPE_FLIPDONTCLEAR = The 'dontclear' flag passed into
Screen('Flip/AsyncFlipBegin', window, when, dontclear), or zero
if none was specified, or -DBL_MAX if unavailable.

IMAGINGPIPE_FLIPVBLSYNCLEVEL = The 'dontsync' flag in
Screen('Flip/AsyncFlipBegin', window, when, dontclear, dontsync),
or zero if none was specified, or -DBL_MAX if unavailable.

Special clients can use these values, in general before the flip
actually executes, to optimize their behavior. The value -DBL_MAX
is special: It would get returned if the requested variable is not
yet available at the time the processing slot of an imaging pipeline
stage/chain gets executed. This would be, e.g., the case if the
processing chain is executed as part of preflip operations, but those
preflip operations are not triggered by a call to Screen('Flip') or
Screen('AsyncFlipBegin') where the info from the user-script is available,
but the preflip ops are triggered by a call to Screen('DrawingFinished'),
where the processing slot would get executed, but usercode can not provide
the actual values to use for those variables. At verbosity 10 this will
additionally trigger proper usage warnings.

-> ImagingPipelineExternalConsumerTest.m fixed and update to exercise
   this new feature.