Screen: Add support for cursor constraining.
The new Screen('ConstrainCursor') function allows to use ClipCursor()
on MS-Windows, and pointer barriers on Linux/X11 to constrain mouse
cursor movements:

On MS-Windows, mouse movement can be constrained to the inside of
a rectangle. On Linux ditto, but also for multiple rectangles, and
one can add arbitrary horizontal or vertical border lines over which
the cursor can't pass. The compile time limit for Linux is currently
1024 lines or 1024/4 rects. On Windows we only have one rect for the
whole desktop session.

Constraints can be added and removed per function, all can be
removed for an onscreen window, and all get removed at window close.

-> Successfully tested on Linux/X11 single-display/single-x-screen
   so far. Test on Windows pending.

-> This function is unsupported/unsupportable on macOS, as the OS
   lacks suitable API.