DrawFormattedText2: Add support for per-word bounding boxes 'wordbounds'.
Like in DrawFormattedText.m, we now optionally return per-word
bounding boxes, e.g., to define AOI's for text reading experiments
(gaze tracking and stuff).

The current code has one problem: It fails if 'transform' is used with
not one transform (scale,translate,rotate,flip), but multiple concatenated
ones. In 8 hours of trying i didn't manage to find the reason, so i give
up for now. Something is very weird in the way the coordinate systems for
these transforms are handled...

-> DrawFormattedText2Demo.m updated to demo per-word bounding boxes.