VRNext: Make Screen('GetImage') from front buffer work with HMDs.
If front buffer snapshot is requested in redirected output mode, do just
that - Read the windowing system frontbuffer of our onscreen window.
Drop the warnings about this being unsupported, and use the windows
'GlobalRect' as reference rectangle for the framebuffer, as that
contains the true onscreen window framebuffer size - does not get
overriden by the VR client.

-> This works fine if the external display client can blit a suitable
image into the onscreen windows backbuffer + OpenGL bufferswap, sourced
from its own display system to provide some approximation of the final
output to the external display device. E.g., for Oculus HMD's this would
be the content of the OpenGL Mirrortexture provided by the VR compositor
with the final output image as rendered to the HMD, just rescaled to fit
the onscreen windows framebuffer.