PsychOculusVR: Add compatibility return arguments for 'PrepareRender'
state.tracked always has +128 flag set to claim successfull display to
the HMD, as we can't detect otherwise.
state.SessionState always fakes a values of 7 for HMD active,
functioning properly, strapped onto users head, no recenter or quit
requests, as we can't detect any of this state.
reqmask 2 returns fake tracked hand positions and orientation of "hands
resting upright in origin", all derived info derived as if this would be
truly the case. handStatus of course reports "not tracked/invalid data",
but we are cautious with regards to users that don't code carefully or
can't read manuals.