Screen('LoadNormalizedGammaTable'): Also return an (empty) oldLUT for 'loadOnNextFlip' == 2.
So far we didn't return an old LUT return arg 1 if 'loadOnNextFlip' was set to 2, as we
don't have anything meaningful to return. However, this makes it a tad more confusing to
get to the 2nd return argument 'success' - which btw. is also pretty meaningless for
anything than 'loadOnNextFlip' == 0, but anyway. We aborted if somebody tried to assign
the 'oldlut' argument, instead of using ~ to skip that assignment.

Now we return an empty matrix as 'oldlut' - problem solved.