PsychImaging: Fix broken 'EnableCLUTMapping' task in functions!
I missed a location where evalin('base', ...) was used to evaluate
PsychHelperCreateRemapCLUT to get to the IMAGINGPIPE_GAMMATABLE
which used to be returned in the runtime base workspace, but now
is returned in the callers workspace for higher performance. The
missing conversion of this user caused a situation where CLUT mapping
only worked if the calling code was in a M-Script (= caller workspace
 equals base workspace), not in a M-Function (caller workspace !=
base workspace). Denis Pelli stumbled over this bug.

-> CLUTMappingBugTest.m (a function!) will verify the bugfix.