ColorCal2: Add support for MS-Windows via virtual COM port communication.
On MS-Windows, no suitable class driver exists to support the ColorCal-II.
Therefore we can't use USB control transfers as on Linux and OSX. However,
access via the generic CDC driver is possible and so a communication channel
over a virtual COM port is exposed.

Implement a 2nd path which uses the virtual COM port via IOPort to talk
to the device. This is probably slower and less reliable, and it only
provides access to a more restricted minimal subset of functions - only
the basics absolutely needed to make the device useful: 'DeviceInfo',
'ZeroCalibration', 'ReadColorMatrix', and 'MeasureXYZ'. Control over the
LED, trigger thresholds, firmware upgrades, raw data readback etc. is not
supported.

This was tested on Linux, enforcing serial port communication, as Linux
supports both USB and serial port access for the ColorCal2.

-> Test under MS-Windows pending.

-> Help text and formatting cleanups as well.