Remove all traces of long dead SerialComm() serial port driver for OSX.
This was a 3rd party binary only mex file for serial port access
under MacOSX. However, it was long superseded by our IOPort driver,
and only available for OSX + 32-Bit Matlab, so it didn't survive
the transition to 64-Bit OSX or Octave.

Now that all dependencies to SerialComm() are removed, remove the
help file, Contents entries, license disclaimer and the XCode
project files for this long dead driver.

R.i.p. and thanks for all the fish.