Added a return parameter for passing the dither API version (or for future use).
In case of AMD graphics cards and Windows, the ATIRadeonperf_Windows utility prints information regarding the dither API version, which now will be picked up by PsychGPUControl.m and passed on to, for example, LoadIdentityClut.m. The return parameter is assigned to empty if not used or ATIRadeonperf_Windows did not provide this information.