Added a new CLUT (gfxhwtype==5) for AMD/Windows.
This CLUT is only used, if ATIRadeonRerf_Windows detected the new dither API (dubbed dither API v2). With a couple of GPU core "R600" cards and graphics driver versions (which expose the new API) and Windows 7 and 10, this table has been shown to be better or as good as any other tables already implemented (sometimes, gfxhwtype==4 would also work).