Added NTP synchronization to NetStation.m (including GetNTP.m, a simple NTP client)
The new NetStation 'GetNTPSynchronize' command implements the EGI N-type
synchronization (eci_NTPClockSynch) method transmitting a self selected
timestamp. Event latencies are then transmitted as difference to this
timestamp.

This is so far only tested for 400 series amps with NetStation 5.3 and
Psychtoolbox on Linux by Urte Roeber. It might possibly also work for
300 series amps (using the NetStation PC as NTP server) and other
NetStation software versions and OS but would require testing.

The implemented solution does NOT depend on NTP synchronized clocks
between the NetStation (amp) and Psychtoolbox machines (as millisecond
precision cannot be achieved without reliable external clock source).
Rather the current NTP server (i.e. amp) time is queried at event time
using the GetNTP client. Event latency is computed in server time as
difference between the synchronization timestamp and the NTP timestamp
requested at event time. Client time and clock sync do not matter.

If the observed network delay exceeds 2 ms a new timestamp is requested
(up to 4 times; never exceeded one time in any of the tests). Event
latency is corrected by half the observed NTP network delay (thus timing
uncertainty is max. 1 ms) and the delay between event time and NTP
request transmission time (in client time; corrected by the clock drift
estimated by regression on the basis of the preceding NTP packages).

We intensively tested the solution with audio and video (photodiode) and
the EGI AV tester box. Precision was in an approximately -/+0.5 ms range
without exception in several thousands of trials and, thus, about 5
times more precise compared to our tests with the existing solution.

Apart from NTP synchronization only some minor cleaning (replaced
returns in nested functions by ends) was done and the NetStation.m
function is expected to be fully backward compatible.

For future reference:

(a) It is essential to use zero for the fraction of seconds part of the
synchronization timestamp. Drift and offset are observed otherwise.
Presumably, this is due to improper handling of the fraction of seconds
part by the EGI NetStation software.

(b) The alternative S-type synchronization method (relative to amp start
time) is broken in several ways: In response to an S-type
synchronization request not amp start time but current amp time is
received. Thus, amp start time has to be entered manually. The received
NTP timestamp's fraction of seconds part is in microsecond and not in
fractions of 2^32 as expected. The actual protocol differs from
documentation. Finally, the NetStation software does not use amp start
time for synchronization internally but some other timestamp, presumably
the time of synchronization request. As this timestamp includes
(uncontrolled) network delays it might introduce timing jitter between
sessions or blocks. This was also practically observed while testing.

Both only applies to the tested NetStation 5.3.0.1 software and might be
different in older or future version. The applied zero in fraction of
seconds workaround should not hurt in future versions (if the protocol
is not finally messed up).