Screen/Linux: Detect lack of OML_sync_control more reliably.
Also check the glXGetExtensions() string before assuming we have
the extension. Our old detection code detects the extension sometimes
when it isn't fully there. Not a big deal, as it would then test it
and find it to fail and therefore disable it, but it leads to printing
a pointless warning about that failure when it isn't surprising that a
non-existent extension would not work.

-> This removes a false warning message when running on the llvmpipe
   software renderer in VM's or on not quite supported SoC's like the
   ASUS Tinkerboard.