ClockRandSeed: Make it more sane again, hopefully.
Undid dhb's warning, as it promotes writing non-portable code
that only works on specific versions of Matlab, and thereby
creates vendor lock-in. Instead rewrote the function to call
equivalent methods to rng('shuffle') in current Octave and older
versions of Matlab, and rng('shuffle') in current Matlab and
future Octave. That was the whole point of having ClockRandSeed()
in the first place afaics: Something that doesn't break each time
that Mathworks incompetent or careless software engineering breaks
backward compatibility. At least i don't want PTB to degrade to
Mathworks or Apples standards if possible.

-> Lightly tested on Octave 3.8, due to lack of suitable Matlab
   version.

-> This must have been broken on Octave btw., as the version()
   check would always report a prehistoric "Matlab version" < 5
   on Octave, given that current Octave has version 4.x < 5.0.