PsychHIDHelpers: Fix crash-inducing wrong fieldNames count.
The white-space cleanup commit "cleaned up whitespace in some of
the PsychHID files " not only cleaned up white-space as it should,
but also sneaked in a wrong change to the fieldCount number in the
PsychAllocOutStructArray() call from within
PsychHIDReturnEventFromEventBuffer(), inducing a hard crash as
soon as one would use KbEventGet(). Fix it.

-> For god's sake, NEVER mix a large scale white-space cleanup with
   functional changes!!! This may be ok if you really know what you
   are doing, and it is small isolated cleanup and everything was
   extensively tested afterwards, but not if you touch every line
   of the file!!

-> Was this EVER properly tested, before the pull request was
   submitted???!