PsychHID/Linux: Fix failure of KbQueues on multi-x-screen setups.
Turns out we were not receiving key events if the mouse pointer
wasn't located on the DefaultScreen() for the X-Server, as we
only enabled delivery of events from the DefaultRootWindow(),
ergo default X-Screen, not from all root windows of all attached
X-Screens.

Fix this embarassing bug by enabling on all X-Screens for the
display connection.