Screen: Hookfunction 'BackendOverrides' allows now to override projection matrices.
This allows to set custom projection matrices to override the standard
ortho matrix setup. This is quite half-baked in that it only applies
to everything but text drawing via drawing plugin, so it doesn't really
work consistently for all scripts -- not for ones involving text.

Anyway, the work is done and has proven pointless for its intended
purpose. Integrate anyway, so the work isn't wasted.