NetStation: Add response check to GetNTP synchronization.
The byte returned by NetStation in response to a GetNTP synchronization
request was not evaluated (as in most other parts of NetStation.m).
Synchronization could silently fail if NetStation software is connected
but not in acquisition mode. The patch adds a check of the return byte.
An info message is printed in case of success and an error in case of
failure.